added api dep. in bootstrap. weather services  and colntroller added

+++ Backend/database/migrations/2024_05_03_081216_create_cities_table.php
@@ -1,4 +1,5 @@
 <?php
+
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -12,8 +13,8 @@ class CreateCitiesTable extends Migration
             $table->string('name');
             $table->string('country');
             $table->string('state')->nullable();
-            $table->double('latitude', 10, 6);
-            $table->double('longitude', 10, 6);
+            $table->decimal('latitude', 10, 7);
+            $table->decimal('longitude', 10, 7);
             $table->timestamps();
         });
     }
@@ -23,4 +24,3 @@ class CreateCitiesTable extends Migration
         Schema::dropIfExists('cities');
     }
 }
-
